requirements.txt
django
djangorestframework
pymongo
django-cors-headers
python-dotenv
nltk
PyPDF2 
python-docx 
docx2txt
bcrypt